Update status messages on service accessibility check
Changed the status messages in service.sh to be more informative. The success message now indicates that the container is "now accessible", to make it clearer that the service came online during the check. If the service is not available, the script will now say that it is "not yet accessible" instead of saying it is "not accessible", to emphasize that the check will continue and the service might come online later.