Update naming and order for more readability in autodns-toggle.sh
The script autodns-toggle.sh was tweaked to improve readability and order of operations. Now, the 'todo' and 'preposition' variables are defined in both branches of the if statement, rather than defining 'todo' later. This makes it clearer that 'todo' and 'preposition' are both used regardless of the condition's outcome. The echo statement was also modified to include the 'preposition' variable, thus providing more accurate instructions for user.